# complete-api-docs.json
{
    "openapi":  "3.0.0",
    "paths":  {
...
                  "/api/cards":  {
                                     "post":  {
                                                  "operationId":  "CardsController_create",
                                                  "summary":  "Create a new card",
                                                  "description":  "Creates a new flashcard with front/back content, optional explanation, examples, and tags",
                                                  "parameters":  [

                                                                 ],
                                                  "requestBody":  {
                                                                      "required":  true,
                                                                      "description":  "Card creation details including front content, back content, explanation, examples, tags, and difficulty level",
                                                                      "content":  {
                                                                                      "application/json":  {
                                                                                                               "schema":  {
                                                                                                                              "$ref":  "#/components/schemas/CreateCardDto"
                                                                                                                          }
                                                                                                           }
                                                                                  }
                                                                  },
                                                  "responses":  {
                                                                    "201":  {
                                                                                "description":  "Card created successfully"
                                                                            },
                                                                    "400":  {
                                                                                "description":  "Invalid card data"
                                                                            },
                                                                    "401":  {
                                                                                "description":  "Unauthorized"
                                                                            }
                                                                },
                                                  "tags":  [
                                                               "Cards"
                                                           ],
                                                  "security":  [
                                                                   {
                                                                       "bearer":  [

                                                                                  ]
                                                                   }
                                                               ]
                                              },
                                     "get":  {
                                                 "operationId":  "CardsController_findAll",
                                                 "summary":  "List cards (filterable, paginated)",
                                                 "description":  "Retrieves a paginated list of user cards with optional filtering by tag and search term",
                                                 "parameters":  [
                                                                    {
                                                                        "name":  "tag",
                                                                        "required":  false,
                                                                        "in":  "query",
                                                                        "description":  "Filter cards by tag",
                                                                        "schema":  {
                                                                                       "example":  "vocabulary",
                                                                                       "type":  "string"
                                                                                   }
                                                                    },
                                                                    {
                                                                        "name":  "search",
                                                                        "required":  false,
                                                                        "in":  "query",
                                                                        "description":  "Search cards by content",
                                                                        "schema":  {
                                                                                       "example":  "serendipity",
                                                                                       "type":  "string"
                                                                                   }
                                                                    },
                                                                    {
                                                                        "name":  "page",
                                                                        "required":  false,
                                                                        "in":  "query",
                                                                        "description":  "Page number for pagination",
                                                                        "schema":  {
                                                                                       "example":  "1",
                                                                                       "type":  "string"
                                                                                   }
                                                                    },
                                                                    {
                                                                        "name":  "pageSize",
                                                                        "required":  false,
                                                                        "in":  "query",
                                                                        "description":  "Number of cards per page",
                                                                        "schema":  {
                                                                                       "example":  "20",
                                                                                       "type":  "string"
                                                                                   }
                                                                    }
                                                                ],
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "Cards retrieved successfully"
                                                                           },
                                                                   "401":  {
                                                                               "description":  "Unauthorized"
                                                                           }
                                                               },
                                                 "tags":  [
                                                              "Cards"
                                                          ],
                                                 "security":  [
                                                                  {
                                                                      "bearer":  [

                                                                                 ]
                                                                  }
                                                              ]
                                             }
                                 },
                  "/api/cards/{id}":  {
                                          "get":  {
                                                      "operationId":  "CardsController_findOne",
                                                      "summary":  "Get a single card",
                                                      "description":  "Retrieves detailed information for a specific card by ID",
                                                      "parameters":  [
                                                                         {
                                                                             "name":  "id",
                                                                             "required":  true,
                                                                             "in":  "path",
                                                                             "description":  "Unique identifier of the card",
                                                                             "schema":  {
                                                                                            "example":  "card-123-uuid-456",
                                                                                            "type":  "string"
                                                                                        }
                                                                         }
                                                                     ],
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "Card retrieved successfully"
                                                                                },
                                                                        "401":  {
                                                                                    "description":  "Unauthorized"
                                                                                },
                                                                        "404":  {
                                                                                    "description":  "Card not found"
                                                                                }
                                                                    },
                                                      "tags":  [
                                                                   "Cards"
                                                               ],
                                                      "security":  [
                                                                       {
                                                                           "bearer":  [

                                                                                      ]
                                                                       }
                                                                   ]
                                                  },
                                          "patch":  {
                                                        "operationId":  "CardsController_update",
                                                        "summary":  "Update a card",
                                                        "description":  "Updates an existing card with new content, tags, or difficulty level",
                                                        "parameters":  [
                                                                           {
                                                                               "name":  "id",
                                                                               "required":  true,
                                                                               "in":  "path",
                                                                               "description":  "Unique identifier of the card to update",
                                                                               "schema":  {
                                                                                              "example":  "card-123-uuid-456",
                                                                                              "type":  "string"
                                                                                          }
                                                                           }
                                                                       ],
                                                        "requestBody":  {
                                                                            "required":  true,
                                                                            "description":  "Card update details including optional front content, back content, explanation, examples, tags, and difficulty level",
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/UpdateCardDto"
                                                                                                                                }
                                                                                                                 }
                                                                                        }
                                                                        },
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "Card updated successfully"
                                                                                  },
                                                                          "400":  {
                                                                                      "description":  "Invalid update data"
                                                                                  },
                                                                          "401":  {
                                                                                      "description":  "Unauthorized"
                                                                                  },
                                                                          "404":  {
                                                                                      "description":  "Card not found"
                                                                                  }
                                                                      },
                                                        "tags":  [
                                                                     "Cards"
                                                                 ],
                                                        "security":  [
                                                                         {
                                                                             "bearer":  [

                                                                                        ]
                                                                         }
                                                                     ]
                                                    },
                                          "delete":  {
                                                         "operationId":  "CardsController_remove",
                                                         "summary":  "Delete a card",
                                                         "description":  "Permanently removes a card from the user\u0027s collection",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "id",
                                                                                "required":  true,
                                                                                "in":  "path",
                                                                                "description":  "Unique identifier of the card to delete",
                                                                                "schema":  {
                                                                                               "example":  "card-123-uuid-456",
                                                                                               "type":  "string"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "Card deleted successfully"
                                                                                   },
                                                                           "401":  {
                                                                                       "description":  "Unauthorized"
                                                                                   },
                                                                           "404":  {
                                                                                       "description":  "Card not found"
                                                                                   }
                                                                       },
                                                         "tags":  [
                                                                      "Cards"
                                                                  ],
                                                         "security":  [
                                                                          {
                                                                              "bearer":  [

                                                                                         ]
                                                                          }
                                                                      ]
                                                     }
                                      },
...
              },
    "info":  {
                 "title":  "Matrus API",
                 "description":  "AI-powered Leitner System API",
                 "version":  "1.0",
                 "contact":  {

                             }
             },
...
    "servers":  [

                ],
    "components":  {
                       "securitySchemes":  {
                                               "bearer":  {
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT",
                                                              "type":  "http"
                                                          }
                                           },
                       "schemas":  {
...
                                       "CreateCardDto":  {
                                                             "type":  "object",
                                                             "properties":  {
                                                                                "front":  {
                                                                                              "type":  "string",
                                                                                              "description":  "Front side content of the flashcard",
                                                                                              "example":  "What is the meaning of \"serendipity\"?"
                                                                                          },
                                                                                "back":  {
                                                                                             "type":  "string",
                                                                                             "description":  "Back side content of the flashcard (answer)",
                                                                                             "example":  "A pleasant surprise; finding something good without looking for it"
                                                                                         },
                                                                                "explanation":  {
                                                                                                    "type":  "string",
                                                                                                    "description":  "Additional explanation for the card content",
                                                                                                    "example":  "Serendipity comes from a Persian fairy tale about princes who made discoveries by accident"
                                                                                                },
                                                                                "exampleSentences":  {
                                                                                                         "description":  "Example sentences demonstrating usage",
                                                                                                         "example":  [
                                                                                                                         "It was pure serendipity that led to their meeting.",
                                                                                                                         "The discovery was a result of serendipity."
                                                                                                                     ],
                                                                                                         "type":  "array",
                                                                                                         "items":  {
                                                                                                                       "type":  "string"
                                                                                                                   }
                                                                                                     },
                                                                                "tags":  {
                                                                                             "description":  "Tags for categorizing the card",
                                                                                             "example":  [
                                                                                                             "vocabulary",
                                                                                                             "english",
                                                                                                             "advanced"
                                                                                                         ],
                                                                                             "type":  "array",
                                                                                             "items":  {
                                                                                                           "type":  "string"
                                                                                                       }
                                                                                         },
                                                                                "media":  {
                                                                                              "type":  "object",
                                                                                              "description":  "Media attachments (images, audio, etc.)"
                                                                                          },
                                                                                "difficulty":  {
                                                                                                   "type":  "number",
                                                                                                   "description":  "Difficulty level from 1 (easy) to 5 (very hard)",
                                                                                                   "example":  3,
                                                                                                   "minimum":  1,
                                                                                                   "maximum":  5
                                                                                               }
                                                                            },
                                                             "required":  [
                                                                              "front",
                                                                              "back"
                                                                          ]
                                                         },
...
                                   }
                   }
}